Rearrange repeated code
