+++ README.md
@@ -34,8 +34,6 @@ HTML, CSS, JvaScript, and Webpack.
 
 ## 🚀 Live Demo <a name="live-demo"></a>
 
-> Add a link to your deployed project.
-
 - [Live Demo Link](https://jorgegoco.github.io/Leaderboard/)
 
 ## 💻 Getting Started <a name="getting-started"></a>
