+++ src/index.css
@@ -27,11 +27,11 @@ button {
   border-radius: 4px;
   background-color: green;
   color: white;
-  padding: 5px;
+  padding: 0 5px;
 }
 
 input {
-  border: 1px solid rgb(230, 230, 230);
+  border: 1px solid rgb(200, 200, 255);
   border-radius: 4px;
   background-color: rgb(218, 232, 241);
   padding: 5px;
